% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_setup_helpers.R
\name{bootstrap_papss_standard_error}
\alias{bootstrap_papss_standard_error}
\title{Boostrap estimate of standard error}
\usage{
bootstrap_papss_standard_error(
  cf,
  pupil_var,
  setup,
  warm_start = T,
  N = 100,
  n = 10.1,
  t_max = 930,
  f = 1/(10^24),
  maxiter_inner = 10000,
  maxiter_outer = 25,
  convergence_tol = 1e-08,
  start_lambda = 0.1,
  should_accum_H = F
)
}
\arguments{
\item{cf}{Coefficients recovered by papss::pupil_solve()}

\item{pupil_var}{Pupil column from aggregated data frame}

\item{setup}{Model setup object returned by papss::pupil_solve()}

\item{warm_start}{Whether or not to re-use cf - if F then coefficients are estimated from scratch for every bootstrap sample}

\item{N}{Number of repetitions for boot-strapping}

\item{n}{Choice for parameter defined by Hoeks & Levelt (number of laters)}

\item{t_max}{Choice for parameter defined by Hoeks & Levelt (response maximum in ms)}

\item{f}{Choice for parameter defined by Wierda et al. (scaling factor)}

\item{maxiter_inner}{Maximum steps taken by inner optimizer}

\item{maxiter_outer}{Maximum steps taken by outer optimizer}

\item{convergence_tol}{Convergence check to terminate early}

\item{start_lambda}{Initial lambda value. Must be > 0 if a penalty should be used! Setting this to 0 and maxiter_outer=1, leads to estimation of an un-penalized additive model, i.e., recovers the traditional NNLS estimate used by Wierda et al. (2012) and Denison et al. (2012).}

\item{should_accum_H}{Whether Hessian should be approximated using BFGS rule or not. If not, then least squares Hessian matrix is used. With the BFGS rule models ended up being much smoother in our simulations. So this should be set to true if under-smoothing is observed. However, the BFGS update is much more costly and takes much more time!}
}
\description{
Bootstrap estimate of the standard error in the pulse weights. Calculation
is based on the "bootstrapping residuals" approach presented in chapter 9
of "An introduction to the Bootstrap" by Efron & Tribshirani. Algorithm 6.1
for calculating standard error estimates based on bootstrap samples
from chapter 6 is used for the actual calculations.
}
