% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_setup_helpers.R
\name{create_constant_term}
\alias{create_constant_term}
\title{Intercept term creation}
\usage{
create_constant_term(time)
}
\arguments{
\item{time}{A numeric vector containing positive time values in ms}
}
\description{
Creates a single intercept term (at the population level).
Can be manipulated via "by" argument as done in mgcv (Wood, 2017).
}
