% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_setup_helpers.R
\name{create_slope_term}
\alias{create_slope_term}
\title{Creates a single slope term (at the population level)}
\usage{
create_slope_term(time_unq, n_cat)
}
\arguments{
\item{time_unq}{A numeric vector containing unique positive time values in ms}

\item{n_cat}{Integer, number of categories for which slope should be repeated.}
}
\description{
Can be manipulated via "by" argument as done in mgcv.
Note: Use of unique values that are just repeated requires all levels
of the category factor (usually subjects) to have the same values on the
variable time.
}
