% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_setup_helpers.R
\name{create_spike_matrix_term}
\alias{create_spike_matrix_term}
\title{Create basis part of the model matrix as described by Wood (2017).
Essentially, this means creating a matrix with a column for each
pulse location and to assign a pupil basis (convolved with that location)
to each column.}
\usage{
create_spike_matrix_term(time, pulse_locations, n, t_max, f)
}
\arguments{
\item{time}{A numeric vector containing positive time values in ms}

\item{pulse_locations}{A numeric vector containing index values of pulse loc.}

\item{n}{Parameter defined by Hoeks & Levelt (number of laters)}

\item{t_max}{Parameter defined by Hoeks & Levelt (response maximum in ms)}

\item{f}{Parameter defined by Wierda et al. (scaling factor)}
}
\description{
See: Wood, S. N. (2017). Generalized Additive Models: An Introduction with R,
Second Edition (2nd ed.). Chapman and Hall/CRC.
}
