% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_setup_helpers.R
\name{DEN_SHARED_NNLS_model_setup}
\alias{DEN_SHARED_NNLS_model_setup}
\title{Creates trainings matrix and penalties for a fully penalized
model inspired by the one used by Denison et al., (2020).
Like by Wierda et al. (2012) separate sets of coefficients are estimated per
subject with the h_basis terms. All those coefficients are again constrained
to be non-negative. Model also includes the intercept terms for each subject that
Denison et al. (2012) introduced. These are not constrained - since their
primary purpose was to account for pupil trajectories that were overall just
containing very negative samples (relative to baseline).}
\usage{
DEN_SHARED_NNLS_model_setup(time, subs, pulse_locations, n, t_max, f)
}
\arguments{
\item{time}{A numeric vector containing positive time values in ms}

\item{subs}{A factor vector containing subject identifiers}

\item{pulse_locations}{A numeric vector containing index values of pulse loc.}

\item{n}{Parameter defined by Hoeks & Levelt (number of laters)}

\item{t_max}{Parameter defined by Hoeks & Levelt (response maximum in ms)}

\item{f}{Parameter defined by Wierda et al. (scaling factor)}
}
\description{
This model differs from the one by Wierda et al. (2012) and Denison et al., (2020)
in that it penalizes the coefficients corresponding to the h_basis terms.
Specifically, it enforces a single penalty term shared by all subjects (e.g., this is
similar to the 'fs' basis in mgcv or can be achieved by using the 'id' keyword
in a 'by' factor smooth.). The form of the penalty expressed on all of the
basis functions is a simple identity matrix. We here also penalize all intercept
terms, again with a single penalty (this time applied to a single matrix though).
}
