% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pupil_sim_helpers.R
\name{get_knots}
\alias{get_knots}
\title{Calculate knot sequence for a B-spline
See Eilers & Marx (2010) for details, the code was
taken from their paper.}
\usage{
get_knots(xl, xr, ndx, bdeg)
}
\arguments{
\item{xl}{minimum of X}

\item{xr}{maximum of X}

\item{ndx}{number of intervals}

\item{bdeg}{degree of basis function (2 for cubic)}
}
\description{
Calculate knot sequence for a B-spline
See Eilers & Marx (2010) for details, the code was
taken from their paper.
}
