\name{papss-package}
\alias{papss-package}
\alias{papss}
\docType{package}
\title{
\packageTitle{papss}
}
\description{
\packageDescription{papss}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{papss}
\packageIndices{papss}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{papss}

Maintainer: \packageMaintainer{papss}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
