% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{solveAM}
\alias{solveAM}
\title{Fits an additive model based on the stable LS solutions discussed in Wood (2011,2017).}
\usage{
solveAM(
  X,
  y,
  initCf,
  constraints,
  lambdaTermFreq,
  lambdaTermDim,
  startIndex,
  maxIter,
  maxIterOptim,
  tol = 0.001
)
}
\description{
Fits an additive model based on the stable LS solutions discussed in Wood (2011,2017).
}
