% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_setup_helpers.R
\name{create_slope_term}
\alias{create_slope_term}
\title{Slope term creation}
\usage{
create_slope_term(time, fact)
}
\arguments{
\item{time}{A numeric vector containing positive time values in ms}

\item{fact}{The factor column from the data-frame passed to papss::pupil_solve()}
}
\description{
Creates a single slope term (at the population level)
Can be manipulated via "by" argument as done in mgcv (Wood, 2017).
Note: Use of unique values that are just repeated requires all levels
of the category factor (usually subjects) to have the same values on the
variable time.
}
