% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_setup_helpers.R
\name{DEN_SHARED_NNLS_model_setup}
\alias{DEN_SHARED_NNLS_model_setup}
\title{Model template based on Denison et al.(2020)'s model investigation}
\usage{
DEN_SHARED_NNLS_model_setup(
  expanded_time,
  expand_by,
  time,
  fact,
  pulse_locations,
  n,
  t_max,
  f
)
}
\arguments{
\item{expanded_time}{A numeric vector containing positive time values in ms, expanded by a certain amount of ms}

\item{expand_by}{Expansion time in ms passed to papss::pupil_solve(expand_by=) divided by sample length in ms}

\item{time}{A numeric vector containing positive time values in ms}

\item{fact}{The factor column from the data-frame passed to papss::pupil_solve()}

\item{pulse_locations}{A numeric vector containing index values of pulse loc.}

\item{n}{Parameter defined by Hoeks & Levelt (number of laters)}

\item{t_max}{Parameter defined by Hoeks & Levelt (response maximum in ms)}

\item{f}{Parameter defined by Wierda et al. (scaling factor)}
}
\description{
Creates trainings matrix and penalties for a fully penalized
version inspired by the models investigated by Denison et al. (2020).
Similarly, separate sets of coefficients are estimated per
subject with the h_basis terms. All those coefficients are again constrained
to be non-negative. The model also includes an intercept/offset term for each
subject as introduced by Denison et al., (2020. These are not constrained.
This model again penalizes
the coefficients corresponding to the h_basis terms. Specifically, it
enforces a single penalty term shared by all subjects (e.g., this is
similar to the 'fs' basis in mgcv). The form of the penalty expressed on all of the
basis functions is a simple identity matrix. We here also penalize all intercept
terms, again with a single penalty.
}
