% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pupil_sim_helpers.R
\name{extract_demand_for_fact}
\alias{extract_demand_for_fact}
\title{Extracts demand curves}
\usage{
extract_demand_for_fact(
  aggr_dat,
  recovered_coef,
  pulse_locations,
  real_locations,
  pulses_in_time,
  expanded_time,
  expanded_by,
  factor_id = "subject",
  n_fact = 10.1,
  t_max_fact = 930,
  f_fact = 1/(10^24)
)
}
\arguments{
\item{aggr_dat}{Aggregated data passed to solver}

\item{recovered_coef}{The coefficients returned by papss::pupil_solve()}

\item{pulse_locations}{The index vector pointing at pulse locations passed to papss::pupil_solve()}

\item{real_locations}{The vector with the time-points at which pulses are assumed passed to papss::pupil_solve()}

\item{pulses_in_time}{Boolean vector indicating which pulse was within the time-window. Returned by papss::pupil_solve()}

\item{expanded_time}{The expanded time series returned by papss::pupil_solve()}

\item{expanded_by}{Expansion time in ms passed to papss::pupil_solve(expand_by=) divided by sample length in ms}

\item{factor_id}{Name of factor column in aggr_dat}

\item{n_fact}{Choice for n parameter}

\item{t_max_fact}{Choice for t_max parameter}

\item{f_fact}{Choice for f parameter}
}
\description{
Extracts demand curves for each level of the factor passed to the pupil_solve
call.
}
