% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pupil_sim_helpers.R
\name{plot_sim_vs_recovered}
\alias{plot_sim_vs_recovered}
\title{Plot simulated pupil data against recovered weights}
\usage{
plot_sim_vs_recovered(
  n_sub,
  aggr_dat,
  sim_obj,
  recovered_coef,
  pulse_locations,
  real_locations
)
}
\arguments{
\item{n_sub}{How many subjects were simulate.}

\item{aggr_dat}{Aggregated simulation data}

\item{sim_obj}{The list returned by additive_pupil_sim}

\item{recovered_coef}{The coefficients returned by papss:pupil_solve()}

\item{pulse_locations}{The index vector pointing at pulse locations passed to papss:pupil_solve()}

\item{real_locations}{The vector with the time-points at which pulses are assumed passed to papss:pupil_solve()}
}
\description{
Plot simulated pupil data against recovered weights
}
