% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_setup_helpers.R
\name{pupil_solve}
\alias{pupil_solve}
\title{The main wrapper function that is also exposed. Fits the desired model.}
\usage{
pupil_solve(
  pulse_locations,
  real_locations,
  data,
  model = "WIER_SHARED",
  n = 10.1,
  t_max = 930,
  f = 1/(10^24),
  maxiter_inner = 10000,
  maxiter_outer = 25,
  convergence_tol = 1e-06
)
}
\description{
The main wrapper function that is also exposed. Fits the desired model.
}
